requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.3
django-cors-headers==4.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
python-telegram-bot==21.7
sniffio==1.3.1
sqlparse==0.5.2
